# src/main/resources/api.yaml
openapi: 3.0.1
info:
  title: OpenAPI definition
  version: v0
servers:
- url: http://localhost:8090
  description: Generated server url
...
paths:
...
  /api/actors/{id}/title:
    get:
      tags:
      - actor-property-reference-controller
      description: get-title-by-actor-Id
      operationId: followPropertyReference-actor-get_1
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            application/hal+json:
              schema:
                $ref: '#/components/schemas/CollectionModelTitle'
            text/uri-list:
              schema:
                type: string
        "404":
          description: Not Found
    put:
      tags:
      - actor-property-reference-controller
      description: update-title-by-actor-Id
      operationId: createPropertyReference-actor-put
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CollectionModelObject'
          application/x-spring-data-compact+json:
            schema:
              $ref: '#/components/schemas/CollectionModelObject'
          text/uri-list:
            schema:
              type: string
        required: true
      responses:
        "200":
          description: OK
          content:
            application/hal+json:
              schema:
                $ref: '#/components/schemas/CollectionModelTitle'
        "201":
          description: Created
          content:
            application/hal+json:
              schema:
                $ref: '#/components/schemas/CollectionModelTitle'
        "204":
          description: No Content
    delete:
      tags:
      - actor-property-reference-controller
      description: delete-title-by-actor-Id
      operationId: deletePropertyReference-actor-delete
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
    patch:
      tags:
      - actor-property-reference-controller
      description: patch-title-by-actor-Id
      operationId: createPropertyReference-actor-patch
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CollectionModelObject'
          application/x-spring-data-compact+json:
            schema:
              $ref: '#/components/schemas/CollectionModelObject'
          text/uri-list:
            schema:
              type: string
        required: true
      responses:
        "200":
          description: OK
          content:
            application/hal+json:
              schema:
                $ref: '#/components/schemas/CollectionModelTitle'
        "204":
          description: No Content
  /api/actors/{id}/title/{propertyId}:
    get:
      tags:
      - actor-property-reference-controller
      description: get-title-by-actor-Id
      operationId: followPropertyReference-actor-get
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      - name: propertyId
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            application/hal+json:
              schema:
                $ref: '#/components/schemas/CollectionModelTitle'
        "404":
          description: Not Found
    delete:
      tags:
      - actor-property-reference-controller
      description: delete-title-by-actor-Id
      operationId: deletePropertyReferenceId-actor-delete
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      - name: propertyId
        in: path
        required: true
        schema:
          type: string
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
...
  /api/categories/{id}/title:
    get:
      tags:
      - category-property-reference-controller
      description: get-title-by-category-Id
      operationId: followPropertyReference-category-get_1
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            application/hal+json:
              schema:
                $ref: '#/components/schemas/CollectionModelTitle'
            text/uri-list:
              schema:
                type: string
        "404":
          description: Not Found
    put:
      tags:
      - category-property-reference-controller
      description: update-title-by-category-Id
      operationId: createPropertyReference-category-put
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CollectionModelObject'
          application/x-spring-data-compact+json:
            schema:
              $ref: '#/components/schemas/CollectionModelObject'
          text/uri-list:
            schema:
              type: string
        required: true
      responses:
        "200":
          description: OK
          content:
            application/hal+json:
              schema:
                $ref: '#/components/schemas/CollectionModelTitle'
        "201":
          description: Created
          content:
            application/hal+json:
              schema:
                $ref: '#/components/schemas/CollectionModelTitle'
        "204":
          description: No Content
    delete:
      tags:
      - category-property-reference-controller
      description: delete-title-by-category-Id
      operationId: deletePropertyReference-category-delete
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
    patch:
      tags:
      - category-property-reference-controller
      description: patch-title-by-category-Id
      operationId: createPropertyReference-category-patch
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CollectionModelObject'
          application/x-spring-data-compact+json:
            schema:
              $ref: '#/components/schemas/CollectionModelObject'
          text/uri-list:
            schema:
              type: string
        required: true
      responses:
        "200":
          description: OK
          content:
            application/hal+json:
              schema:
                $ref: '#/components/schemas/CollectionModelTitle'
        "204":
          description: No Content
  /api/categories/{id}/title/{propertyId}:
    get:
      tags:
      - category-property-reference-controller
      description: get-title-by-category-Id
      operationId: followPropertyReference-category-get
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      - name: propertyId
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            application/hal+json:
              schema:
                $ref: '#/components/schemas/CollectionModelTitle'
        "404":
          description: Not Found
    delete:
      tags:
      - category-property-reference-controller
      description: delete-title-by-category-Id
      operationId: deletePropertyReferenceId-category-delete
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      - name: propertyId
        in: path
        required: true
        schema:
          type: string
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
...
  /api/directors/{id}/title:
    get:
      tags:
      - director-property-reference-controller
      description: get-title-by-director-Id
      operationId: followPropertyReference-director-get_1
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            application/hal+json:
              schema:
                $ref: '#/components/schemas/CollectionModelTitle'
            text/uri-list:
              schema:
                type: string
        "404":
          description: Not Found
    put:
      tags:
      - director-property-reference-controller
      description: update-title-by-director-Id
      operationId: createPropertyReference-director-put
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CollectionModelObject'
          application/x-spring-data-compact+json:
            schema:
              $ref: '#/components/schemas/CollectionModelObject'
          text/uri-list:
            schema:
              type: string
        required: true
      responses:
        "200":
          description: OK
          content:
            application/hal+json:
              schema:
                $ref: '#/components/schemas/CollectionModelTitle'
        "201":
          description: Created
          content:
            application/hal+json:
              schema:
                $ref: '#/components/schemas/CollectionModelTitle'
        "204":
          description: No Content
    delete:
      tags:
      - director-property-reference-controller
      description: delete-title-by-director-Id
      operationId: deletePropertyReference-director-delete
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
    patch:
      tags:
      - director-property-reference-controller
      description: patch-title-by-director-Id
      operationId: createPropertyReference-director-patch
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CollectionModelObject'
          application/x-spring-data-compact+json:
            schema:
              $ref: '#/components/schemas/CollectionModelObject'
          text/uri-list:
            schema:
              type: string
        required: true
      responses:
        "200":
          description: OK
          content:
            application/hal+json:
              schema:
                $ref: '#/components/schemas/CollectionModelTitle'
        "204":
          description: No Content
  /api/directors/{id}/title/{propertyId}:
    get:
      tags:
      - director-property-reference-controller
      description: get-title-by-director-Id
      operationId: followPropertyReference-director-get
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      - name: propertyId
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            application/hal+json:
              schema:
                $ref: '#/components/schemas/CollectionModelTitle'
        "404":
          description: Not Found
    delete:
      tags:
      - director-property-reference-controller
      description: delete-title-by-director-Id
      operationId: deletePropertyReferenceId-director-delete
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      - name: propertyId
        in: path
        required: true
        schema:
          type: string
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
...
  /api/titles:
    get:
      tags:
      - title-entity-controller
      description: get-title
      operationId: getCollectionResource-title-get_1
      parameters:
      - name: page
        in: query
        description: Zero-based page index (0..N)
        required: false
        schema:
          minimum: 0
          type: integer
          default: 0
      - name: size
        in: query
        description: The size of the page to be returned
        required: false
        schema:
          minimum: 1
          type: integer
          default: 20
      - name: sort
        in: query
        description: "Sorting criteria in the format: property,(asc|desc). Default\
          \ sort order is ascending. Multiple sort criteria are supported."
        required: false
        schema:
          type: array
          items:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/hal+json:
              schema:
                $ref: '#/components/schemas/PagedModelEntityModelTitle'
            application/x-spring-data-compact+json:
              schema:
                $ref: '#/components/schemas/PagedModelEntityModelTitle'
            text/uri-list:
              schema:
                type: string
    post:
      tags:
      - title-entity-controller
      description: create-title
      operationId: postCollectionResource-title-post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TitleRequestBody'
        required: true
      responses:
        "201":
          description: Created
          content:
            application/hal+json:
              schema:
                $ref: '#/components/schemas/EntityModelTitle'
...
  /api/titles/{id}:
    get:
      tags:
      - title-entity-controller
      description: get-title
      operationId: getItemResource-title-get
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            application/hal+json:
              schema:
                $ref: '#/components/schemas/EntityModelTitle'
        "404":
          description: Not Found
    put:
      tags:
      - title-entity-controller
      description: update-title
      operationId: putItemResource-title-put
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TitleRequestBody'
        required: true
      responses:
        "200":
          description: OK
          content:
            application/hal+json:
              schema:
                $ref: '#/components/schemas/EntityModelTitle'
        "201":
          description: Created
          content:
            application/hal+json:
              schema:
                $ref: '#/components/schemas/EntityModelTitle'
        "204":
          description: No Content
    delete:
      tags:
      - title-entity-controller
      description: delete-title
      operationId: deleteItemResource-title-delete
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
    patch:
      tags:
      - title-entity-controller
      description: patch-title
      operationId: patchItemResource-title-patch
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TitleRequestBody'
        required: true
      responses:
        "200":
          description: OK
          content:
            application/hal+json:
              schema:
                $ref: '#/components/schemas/EntityModelTitle'
        "204":
          description: No Content
  /api/titles/{id}/actor:
    get:
      tags:
      - title-property-reference-controller
      description: get-actor-by-title-Id
      operationId: followPropertyReference-title-get_1
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            application/hal+json:
              schema:
                $ref: '#/components/schemas/CollectionModelActor'
            text/uri-list:
              schema:
                type: string
        "404":
          description: Not Found
    put:
      tags:
      - title-property-reference-controller
      description: update-actor-by-title-Id
      operationId: createPropertyReference-title-put
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CollectionModelObject'
          application/x-spring-data-compact+json:
            schema:
              $ref: '#/components/schemas/CollectionModelObject'
          text/uri-list:
            schema:
              type: string
        required: true
      responses:
        "200":
          description: OK
          content:
            application/hal+json:
              schema:
                $ref: '#/components/schemas/CollectionModelActor'
        "201":
          description: Created
          content:
            application/hal+json:
              schema:
                $ref: '#/components/schemas/CollectionModelActor'
        "204":
          description: No Content
    delete:
      tags:
      - title-property-reference-controller
      description: delete-actor-by-title-Id
      operationId: deletePropertyReference-title-delete
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
    patch:
      tags:
      - title-property-reference-controller
      description: patch-actor-by-title-Id
      operationId: createPropertyReference-title-patch
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CollectionModelObject'
          application/x-spring-data-compact+json:
            schema:
              $ref: '#/components/schemas/CollectionModelObject'
          text/uri-list:
            schema:
              type: string
        required: true
      responses:
        "200":
          description: OK
          content:
            application/hal+json:
              schema:
                $ref: '#/components/schemas/CollectionModelActor'
        "204":
          description: No Content
...
  /api/titles/{id}/category:
    get:
      tags:
      - title-property-reference-controller
      description: get-category-by-title-Id
      operationId: followPropertyReference-title-get_2_1
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            application/hal+json:
              schema:
                $ref: '#/components/schemas/CollectionModelCategory'
            text/uri-list:
              schema:
                type: string
        "404":
          description: Not Found
    put:
      tags:
      - title-property-reference-controller
      description: update-category-by-title-Id
      operationId: createPropertyReference-title-put_1
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CollectionModelObject'
          application/x-spring-data-compact+json:
            schema:
              $ref: '#/components/schemas/CollectionModelObject'
          text/uri-list:
            schema:
              type: string
        required: true
      responses:
        "200":
          description: OK
          content:
            application/hal+json:
              schema:
                $ref: '#/components/schemas/CollectionModelCategory'
        "201":
          description: Created
          content:
            application/hal+json:
              schema:
                $ref: '#/components/schemas/CollectionModelCategory'
        "204":
          description: No Content
    delete:
      tags:
      - title-property-reference-controller
      description: delete-category-by-title-Id
      operationId: deletePropertyReference-title-delete_1
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
    patch:
      tags:
      - title-property-reference-controller
      description: patch-category-by-title-Id
      operationId: createPropertyReference-title-patch_1
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CollectionModelObject'
          application/x-spring-data-compact+json:
            schema:
              $ref: '#/components/schemas/CollectionModelObject'
          text/uri-list:
            schema:
              type: string
        required: true
      responses:
        "200":
          description: OK
          content:
            application/hal+json:
              schema:
                $ref: '#/components/schemas/CollectionModelCategory'
        "204":
          description: No Content
...
  /api/titles/{id}/director:
    get:
      tags:
      - title-property-reference-controller
      description: get-director-by-title-Id
      operationId: followPropertyReference-title-get_3_1
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            application/hal+json:
              schema:
                $ref: '#/components/schemas/CollectionModelDirector'
            text/uri-list:
              schema:
                type: string
        "404":
          description: Not Found
    put:
      tags:
      - title-property-reference-controller
      description: update-director-by-title-Id
      operationId: createPropertyReference-title-put_2
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CollectionModelObject'
          application/x-spring-data-compact+json:
            schema:
              $ref: '#/components/schemas/CollectionModelObject'
          text/uri-list:
            schema:
              type: string
        required: true
      responses:
        "200":
          description: OK
          content:
            application/hal+json:
              schema:
                $ref: '#/components/schemas/CollectionModelDirector'
        "201":
          description: Created
          content:
            application/hal+json:
              schema:
                $ref: '#/components/schemas/CollectionModelDirector'
        "204":
          description: No Content
    delete:
      tags:
      - title-property-reference-controller
      description: delete-director-by-title-Id
      operationId: deletePropertyReference-title-delete_2
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
    patch:
      tags:
      - title-property-reference-controller
      description: patch-director-by-title-Id
      operationId: createPropertyReference-title-patch_2
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CollectionModelObject'
          application/x-spring-data-compact+json:
            schema:
              $ref: '#/components/schemas/CollectionModelObject'
          text/uri-list:
            schema:
              type: string
        required: true
      responses:
        "200":
          description: OK
          content:
            application/hal+json:
              schema:
                $ref: '#/components/schemas/CollectionModelDirector'
        "204":
          description: No Content
...
components:
  schemas:
...
    Link:
      type: object
      properties:
        href:
          type: string
        hreflang:
          type: string
        title:
          type: string
        type:
          type: string
        deprecation:
          type: string
        profile:
          type: string
        name:
          type: string
        templated:
          type: boolean
...
    EntityModelTitle:
      required:
      - dateAdded
      - duration
      - name
      - rating
      - releaseYear
      type: object
      properties:
        name:
          type: string
        dateAdded:
          type: string
        releaseYear:
          type: string
        rating:
          type: string
        duration:
          type: string
        description:
          type: string
        userRating:
          maximum: 10
          minimum: 0
          type: number
          format: float
        numRatings:
          minimum: 0
          type: integer
          format: int32
        _links:
          type: array
          items:
            $ref: '#/components/schemas/Link'
    PageMetadata:
      type: object
      properties:
        size:
          type: integer
          format: int64
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
          format: int64
        number:
          type: integer
          format: int64
    PagedModelEntityModelTitle:
      type: object
      properties:
        _embedded:
          type: object
          properties:
            titles:
              type: array
              items:
                $ref: '#/components/schemas/EntityModelTitle'
        _links:
          type: array
          items:
            $ref: '#/components/schemas/Link'
        page:
          $ref: '#/components/schemas/PageMetadata'
    CollectionModelDirector:
      type: object
      properties:
        _embedded:
          type: object
          properties:
            directors:
              type: array
              items:
                $ref: '#/components/schemas/DirectorResponse'
        _links:
          type: array
          items:
            $ref: '#/components/schemas/Link'
    CollectionModelObject:
      type: object
      properties:
        _embedded:
          type: object
          properties:
            objects:
              type: array
              items:
                type: object
        _links:
          type: array
          items:
            $ref: '#/components/schemas/Link'
    CollectionModelCategory:
      type: object
      properties:
        _embedded:
          type: object
          properties:
            categories:
              type: array
              items:
                $ref: '#/components/schemas/CategoryResponse'
        _links:
          type: array
          items:
            $ref: '#/components/schemas/Link'
    CollectionModelActor:
      type: object
      properties:
        _embedded:
          type: object
          properties:
            actors:
              type: array
              items:
                $ref: '#/components/schemas/ActorResponse'
        _links:
          type: array
          items:
            $ref: '#/components/schemas/Link'
...
    CollectionModelTitle:
      type: object
      properties:
        _embedded:
          type: object
          properties:
            titles:
              type: array
              items:
                $ref: '#/components/schemas/TitleResponse'
        _links:
          type: array
          items:
            $ref: '#/components/schemas/Link'
...
    TitleRequestBody:
      required:
      - dateAdded
      - duration
      - name
      - rating
      - releaseYear
      type: object
      properties:
        id:
          type: integer
          format: int32
        name:
          type: string
        director:
          type: array
          items:
            type: string
        actor:
          type: array
          items:
            type: string
        category:
          type: array
          items:
            type: string
        dateAdded:
          type: string
        releaseYear:
          type: string
        rating:
          type: string
        duration:
          type: string
        description:
          type: string
        userRating:
          maximum: 10
          minimum: 0
          type: number
          format: float
        numRatings:
          minimum: 0
          type: integer
          format: int32
    ActorResponse:
      required:
      - name
      type: object
      properties:
        name:
          type: string
    CategoryResponse:
      required:
      - name
      type: object
      properties:
        name:
          type: string
    DirectorResponse:
      required:
      - name
      type: object
      properties:
        name:
          type: string
    TitleResponse:
      required:
      - dateAdded
      - duration
      - name
      - rating
      - releaseYear
      type: object
      properties:
        name:
          type: string
        dateAdded:
          type: string
        releaseYear:
          type: string
        rating:
          type: string
        duration:
          type: string
        description:
          type: string
        userRating:
          maximum: 10
          minimum: 0
          type: number
          format: float
        numRatings:
          minimum: 0
          type: integer
          format: int32
  securitySchemes:
    netflixAuth:
      type: http
      in: header
      scheme: basic